requirements.txt
pdf2image==1.17.0
Pillow==10.0.1